refactor: update exports for move function to call vis without react from index.js to cellpop.js
